Login and registration prototype
You can now register and login!

:moyai:
