+++ templates/main/test.html
@@ -4,7 +4,21 @@
 {% block title %}Sexy{% endblock %}
 
 {% block content %}
-<h1>Hi</h1>
+<h2>
+Account type: {{ accountType }}
+</h2>
 
+<h2>
+Superuser?: {{ superUser }}
+</h2>
 
+<!-- <ol>
+    <li>Username: {{username}}</li>
+    <li>First name: {{firstname}}</li>
+</ol> -->
+<img style="width:100px;height:auto;object-fit:cover" id="roomFormImage" src="{{picture.url}}"
+                alt="your image" />
+<h2>Img src: {{picture.url}}</h2>
+<H3>{{userid}}</H3>
+<H3>{{username}}</H3>
 {% endblock %}
